fix api

+++ src/LambdaTasks/MathOperation.java
@@ -1,5 +1,5 @@
 package LambdaTasks;
 @FunctionalInterface
 public interface MathOperation {
-    double math(int a,int b);
+    int math(int a,int b);
 }
